fix: remove carbon compat and fix tests
