+++ webapp/client/apps/orchestration-cluster-webapp/package.json
@@ -53,7 +53,6 @@
 		"i18next": "26.3.6",
 		"i18next-browser-languagedetector": "8.2.1",
 		"lodash": "4.18.1",
-		"lucide-react": "1.25.0",
 		"mobx": "6.16.1",
 		"mobx-react-lite": "4.1.1",
 		"monaco-editor": "0.53.0",
