--- webapp/client/apps/orchestration-cluster-webapp/src/routes/shadcn/tasklist/-login.test.tsx
@@ -1,76 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-import {it} from '#/vitest-modules/test-extend';
-import {renderWithRouter} from '#/vitest-modules/render-with-router';
-import {describe, expect, vi} from 'vitest';
-import {userEvent} from 'vitest/browser';
-import {Route} from './login';
-
-describe('<TasklistLoginPage />', () => {
-	it('should have the correct copyright notice', async () => {
-		vi.useFakeTimers();
-		const mockYear = 1984;
-		vi.setSystemTime(new Date(mockYear, 0));
-
-		const screen = await renderWithRouter(Route.options.component!, {path: '/shadcn/tasklist/login'});
-
-		await expect
-			.element(screen.getByText(`© Camunda Services GmbH ${mockYear}. All rights reserved. | 0.0.0`))
-			.toBeVisible();
-		vi.useRealTimers();
-	});
-
-	it('should not allow the form to be submitted with empty fields', async () => {
-		const screen = await renderWithRouter(Route.options.component!, {path: '/shadcn/tasklist/login'});
-
-		await userEvent.click(screen.getByRole('button', {name: /login/i}));
-
-		// carbon-compat's TextInput (@camunda/design-system 0.45.0) renders the
-		// error text as a plain <span> with no aria-errormessage linking it to
-		// the input, unlike Carbon's own TextInput used on the Carbon login
-		// page. That's a known upstream gap, not a defect in this page —
-		// tracked at https://github.com/camunda/design-system/issues/483 —
-		// so the username field is asserted via visible text + validity
-		// rather than toHaveAccessibleErrorMessage. The password field is
-		// our own composed component, so it does wire aria-errormessage
-		// itself and keeps the stricter assertion.
-		await expect.element(screen.getByText(/username is required/i)).toBeVisible();
-		await expect.element(screen.getByLabelText(/username/i)).toBeInvalid();
-		await expect.element(screen.getByLabelText(/^password$/i)).toHaveAccessibleErrorMessage(/password is required/i);
-		await expect.element(screen.getByLabelText(/^password$/i)).toBeInvalid();
-
-		await userEvent.fill(screen.getByLabelText(/username/i), 'demo');
-		await userEvent.click(screen.getByRole('button', {name: /login/i}));
-
-		await expect.element(screen.getByLabelText(/username/i)).toBeValid();
-		await expect.element(screen.getByLabelText(/^password$/i)).toHaveAccessibleErrorMessage(/password is required/i);
-		await expect.element(screen.getByLabelText(/^password$/i)).toBeInvalid();
-
-		await userEvent.fill(screen.getByLabelText(/username/i), '');
-		await userEvent.fill(screen.getByLabelText(/^password$/i), 'demo');
-		await userEvent.click(screen.getByRole('button', {name: /login/i}));
-
-		await expect.element(screen.getByLabelText(/^password$/i)).toBeValid();
-		await expect.element(screen.getByText(/username is required/i)).toBeVisible();
-		await expect.element(screen.getByLabelText(/username/i)).toBeInvalid();
-	});
-
-	it('should toggle password visibility', async () => {
-		const screen = await renderWithRouter(Route.options.component!, {path: '/shadcn/tasklist/login'});
-
-		const passwordField = screen.getByLabelText(/^password$/i);
-		await expect.element(passwordField).toHaveAttribute('type', 'password');
-
-		await userEvent.click(screen.getByRole('button', {name: /show password/i}));
-		await expect.element(passwordField).toHaveAttribute('type', 'text');
-
-		await userEvent.click(screen.getByRole('button', {name: /hide password/i}));
-		await expect.element(passwordField).toHaveAttribute('type', 'password');
-	});
-});
