--- webapp/client/apps/orchestration-cluster-webapp/src/tasklist/pages/TasklistLoginPage.module.scss
@@ -1,99 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-.page {
-	display: flex;
-	min-height: 100%;
-	flex-direction: column;
-	align-items: center;
-	padding: calc(var(--spacing) * 4);
-	background: var(--background);
-}
-
-.content {
-	display: flex;
-	flex: 1;
-	width: 100%;
-	align-items: center;
-	justify-content: center;
-}
-
-.formWrapper {
-	width: 100%;
-	max-width: 26rem;
-}
-
-.cardContent {
-	display: flex;
-	flex-direction: column;
-	gap: calc(var(--spacing) * 4);
-}
-
-.logo {
-	display: flex;
-	justify-content: center;
-	padding-top: calc(var(--spacing) * 2);
-}
-
-.logoMark {
-	width: calc(var(--spacing) * 12);
-	height: calc(var(--spacing) * 12);
-}
-
-.title {
-	margin-bottom: calc(var(--spacing) * 2);
-	padding-bottom: calc(var(--spacing) * 2);
-	text-align: center;
-}
-
-.field {
-	display: flex;
-	flex-direction: column;
-}
-
-.button {
-	width: 100%;
-	margin-top: calc(var(--spacing) * 2);
-}
-
-.passwordField {
-	display: flex;
-	flex-direction: column;
-	gap: calc(var(--spacing) * 1.5);
-}
-
-.passwordInputWrapper {
-	position: relative;
-	display: flex;
-}
-
-.passwordInput {
-	padding-right: calc(var(--spacing) * 9);
-}
-
-.passwordToggle {
-	position: absolute;
-	top: 0;
-	right: 0;
-	color: var(--neutral-foreground-subtle);
-}
-
-.passwordToggle:hover {
-	color: var(--foreground);
-}
-
-.fieldError {
-	margin: 0;
-	color: var(--danger-foreground-strong);
-}
-
-.copyrightNotice {
-	display: block;
-	padding: calc(var(--spacing) * 4) 0;
-	text-align: center;
-}
